openapi-synthesize/pyproject.toml
[project]
name = "openapi-tools"
version = "1.0"
description = "openapi tooling"
authors = [
    {name = "Wazo Authors", email = "dev@wazo.io"}
]
license = { text = "GPL-3.0-or-later" }
requires-python = ">=3.9"
dependencies = [
    "xivo @ https://github.com/wazo-platform/xivo-lib-python/archive/master.zip",
    "pyyaml==5.3.1",
    "setuptools"
]

[project.urls]
Homepage = "https://wazo.io"
Repository = "https://github.com/wazo-platform/wazo-tools.git"

[project.scripts]
openapi-synthesize = "openapi_synthesize:main"

[build-system]
requires = ["setuptools", "setuptools-scm", "wheel", "toml"]
build-backend = "setuptools.build_meta"

[tool.mypy]
python_version = "3.10"
show_error_codes = true
check_untyped_defs = true
warn_unused_configs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_unused_ignores = true
strict_equality = true
no_warn_no_return = true

[tool.flake8]
show-source = true
max-line-length = 99
ignore = [
  "E203", # whitespace before ':'
  "W503", # line break before binary operator
]
exclude = [
    ".tox",
    ".eggs",
]

[tool.isort]
py_version=39
profile = "black"

